fix: move flattenStylesheets to shared
